OT146-57 | rename 'storage' to 'cloud'
